requirements.txt
flask
pytesseract
pillow
gunicorn